+++ src/ray/core_worker/task_interface.cc
@@ -7,147 +7,6 @@
 
 namespace ray {
 
-ActorHandle::ActorHandle(
-    const class ActorID &actor_id, const class ActorHandleID &actor_handle_id,
-    const class JobID &job_id, const ObjectID &initial_cursor,
-    const Language actor_language, bool is_direct_call,
-    const std::vector<std::string> &actor_creation_task_function_descriptor) {
-  inner_.set_actor_id(actor_id.Data(), actor_id.Size());
-  inner_.set_actor_handle_id(actor_handle_id.Data(), actor_handle_id.Size());
-  inner_.set_job_id(job_id.Data(), job_id.Size());
-  inner_.set_actor_language(actor_language);
-  *inner_.mutable_actor_creation_task_function_descriptor() = {
-      actor_creation_task_function_descriptor.begin(),
-      actor_creation_task_function_descriptor.end()};
-  const auto &actor_creation_task_id = TaskID::ForActorCreationTask(actor_id);
-  const auto &actor_creation_dummy_object_id =
-      ObjectID::ForTaskReturn(actor_creation_task_id, /*index=*/1, /*transport_type=*/0);
-  inner_.set_actor_cursor(actor_creation_dummy_object_id.Data(),
-                          actor_creation_dummy_object_id.Size());
-  inner_.set_is_direct_call(is_direct_call);
-  IncrementTaskCounter();
-  SetActorCursor(initial_cursor);
-}
-
-ActorHandle::ActorHandle(ActorHandle &parent, bool in_band) {
-  std::unique_lock<std::mutex> guard(parent.mutex_);
-  inner_ = parent.inner_;
-  class ActorHandleID new_actor_handle_id;
-  if (in_band) {
-    new_actor_handle_id = ComputeForkedActorHandleId(
-        ActorHandleID::FromBinary(parent.inner_.actor_handle_id()),
-        parent.IncrementNumForks());
-    // Notify the backend to expect this new actor handle. The backend will
-    // not release the cursor for any new handles until the first task for
-    // each of the new handles is submitted.
-    // NOTE(swang): There is currently no garbage collection for actor
-    // handles until the actor itself is removed.
-    parent.new_actor_handles_.push_back(new_actor_handle_id);
-  } else {
-    // If this serialization is happening out-of-band, we set the actor handle ID.
-    // to nil to signal that it should be computed when the handle is deserialized.
-    new_actor_handle_id = ActorHandleID::Nil();
-    // The execution dependency for a pickled actor handle is never safe
-    // to release, since it could be unpickled and submit another
-    // dependent task at any time. Therefore, we notify the backend of a
-    // random handle ID that will never actually be used.
-    parent.new_actor_handles_.push_back(ActorHandleID::FromRandom());
-  }
-  guard.unlock();
-
-  inner_.set_actor_handle_id(new_actor_handle_id.Data(), new_actor_handle_id.Size());
-  inner_.set_task_counter(0);
-  inner_.set_num_forks(0);
-}
-
-ActorHandle::ActorHandle(const std::string &serialized, const TaskID &current_task_id) {
-  inner_.ParseFromString(serialized);
-  // If the actor handle ID is nil, this serialized handle was created by an out-of-band
-  // mechanism (see fork constructor above), so we compute a new actor handle ID.
-  // TODO(pcm): This still leads to a lot of actor handles being
-  // created, there should be a better way to handle pickled
-  // actor handles.
-  // TODO(swang): Unpickling the same actor handle twice in the same
-  // task will break the application, and unpickling it twice in the
-  // same actor is likely a performance bug. We should consider
-  // logging a warning in these cases.
-  if (ActorHandleID::FromBinary(inner_.actor_handle_id()).IsNil()) {
-    const class ActorHandleID new_actor_handle_id = ComputeOutOfBandActorHandleId(
-        ActorHandleID::FromBinary(inner_.actor_handle_id()), current_task_id);
-    inner_.set_actor_handle_id(new_actor_handle_id.Data(), new_actor_handle_id.Size());
-  }
-}
-
-ActorID ActorHandle::ActorID() const { return ActorID::FromBinary(inner_.actor_id()); };
-
-ActorHandleID ActorHandle::ActorHandleID() const {
-  return ActorHandleID::FromBinary(inner_.actor_handle_id());
-};
-
-JobID ActorHandle::JobID() const { return JobID::FromBinary(inner_.job_id()); };
-
-Language ActorHandle::ActorLanguage() const { return inner_.actor_language(); };
-
-std::vector<std::string> ActorHandle::ActorCreationTaskFunctionDescriptor() const {
-  return VectorFromProtobuf(inner_.actor_creation_task_function_descriptor());
-};
-
-void ActorHandle::SetActorTaskSpec(TaskSpecBuilder &builder,
-                                   const TaskTransportType transport_type,
-                                   const ObjectID new_cursor) {
-  std::unique_lock<std::mutex> guard(mutex_);
-  // Build actor task spec.
-  const TaskID actor_creation_task_id = TaskID::ForActorCreationTask(ActorID());
-  const ObjectID actor_creation_dummy_object_id =
-      ObjectID::ForTaskReturn(actor_creation_task_id, /*index=*/1,
-                              /*transport_type=*/static_cast<int>(transport_type));
-  builder.SetActorTaskSpec(ActorID(), ActorHandleID(), actor_creation_dummy_object_id,
-                           /*previous_actor_task_dummy_object_id=*/ActorCursor(),
-                           IncrementTaskCounter(), NewActorHandles());
-
-  SetActorCursor(new_cursor);
-  ClearNewActorHandles();
-}
-
-ObjectID ActorHandle::ActorCursor() const {
-  return ObjectID::FromBinary(inner_.actor_cursor());
-};
-
-int64_t ActorHandle::TaskCounter() const { return inner_.task_counter(); };
-
-int64_t ActorHandle::IncrementTaskCounter() {
-  int64_t old = inner_.task_counter();
-  inner_.set_task_counter(old + 1);
-  return old;
-}
-
-int64_t ActorHandle::NumForks() const { return inner_.num_forks(); };
-
-int64_t ActorHandle::IncrementNumForks() {
-  int64_t old = inner_.num_forks();
-  inner_.set_num_forks(old + 1);
-  return old;
-}
-
-bool ActorHandle::IsDirectCallActor() const { return inner_.is_direct_call(); }
-
-void ActorHandle::Serialize(std::string *output) {
-  std::unique_lock<std::mutex> guard(mutex_);
-  inner_.SerializeToString(output);
-}
-
-ActorHandle::ActorHandle() {}
-
-void ActorHandle::SetActorCursor(const ObjectID &actor_cursor) {
-  inner_.set_actor_cursor(actor_cursor.Binary());
-};
-
-std::vector<ActorHandleID> ActorHandle::NewActorHandles() const {
-  return new_actor_handles_;
-}
-
-void ActorHandle::ClearNewActorHandles() { new_actor_handles_.clear(); }
-
 CoreWorkerTaskInterface::CoreWorkerTaskInterface(
     WorkerContext &worker_context, std::unique_ptr<RayletClient> &raylet_client,
     CoreWorkerObjectInterface &object_interface, boost::asio::io_service &io_service,
@@ -255,9 +114,9 @@ Status CoreWorkerTaskInterface::SubmitActorTask(ActorHandle &actor_handle,
   const TaskID actor_task_id = TaskID::ForActorTask(
       worker_context_.GetCurrentJobID(), worker_context_.GetCurrentTaskID(),
       next_task_index, actor_handle.ActorID());
-  BuildCommonTaskSpec(builder, actor_handle.JobID(), actor_task_id, next_task_index,
-                      function, args, num_returns, task_options.resources, {},
-                      transport_type, return_ids);
+  BuildCommonTaskSpec(builder, actor_handle.CreationJobID(), actor_task_id,
+                      next_task_index, function, args, num_returns,
+                      task_options.resources, {}, transport_type, return_ids);
 
   const ObjectID new_cursor = return_ids->back();
   actor_handle.SetActorTaskSpec(builder, transport_type, new_cursor);
@@ -265,7 +124,7 @@ Status CoreWorkerTaskInterface::SubmitActorTask(ActorHandle &actor_handle,
   // Submit task.
   auto status = task_submitters_[transport_type]->SubmitTask(builder.Build());
   // Remove cursor from return ids.
-  (*return_ids).pop_back();
+  return_ids->pop_back();
 
   return status;
 }
