+++ src/ray/core_worker/test/core_worker_test.cc
@@ -780,26 +780,26 @@ TEST_F(ZeroNodeTest, TestActorHandle) {
 
   // Test in-band forking logic.
   ActorHandle forkedHandle1(parent, true);
-  ASSERT_EQ(1, parent.NumForks());
+  ASSERT_EQ(1, parent.num_forks_);
   ASSERT_EQ(parent.ActorID(), forkedHandle1.ActorID());
   ASSERT_NE(parent.ActorHandleID(), forkedHandle1.ActorHandleID());
   ASSERT_EQ(parent.ActorLanguage(), forkedHandle1.ActorLanguage());
   ASSERT_EQ(parent.ActorCreationTaskFunctionDescriptor(),
             forkedHandle1.ActorCreationTaskFunctionDescriptor());
   ASSERT_EQ(parent.ActorCursor(), forkedHandle1.ActorCursor());
-  ASSERT_EQ(0, forkedHandle1.TaskCounter());
-  ASSERT_EQ(0, forkedHandle1.NumForks());
-  ASSERT_EQ(parent.NewActorHandles().size(), 1);
-  ASSERT_EQ(parent.NewActorHandles().back(), forkedHandle1.ActorHandleID());
-  parent.ClearNewActorHandles();
+  ASSERT_EQ(0, forkedHandle1.task_counter_);
+  ASSERT_EQ(0, forkedHandle1.num_forks_);
+  ASSERT_EQ(parent.new_actor_handles_.size(), 1);
+  ASSERT_EQ(parent.new_actor_handles_.back(), forkedHandle1.ActorHandleID());
+  parent.new_actor_handles_.clear();
 
   ActorHandle forkedHandle2(parent, true);
-  ASSERT_EQ(2, parent.NumForks());
-  ASSERT_EQ(0, forkedHandle2.TaskCounter());
-  ASSERT_EQ(0, forkedHandle2.NumForks());
-  ASSERT_EQ(parent.NewActorHandles().size(), 1);
-  ASSERT_EQ(parent.NewActorHandles().back(), forkedHandle2.ActorHandleID());
-  parent.ClearNewActorHandles();
+  ASSERT_EQ(2, parent.num_forks_);
+  ASSERT_EQ(0, forkedHandle2.task_counter_);
+  ASSERT_EQ(0, forkedHandle2.num_forks_);
+  ASSERT_EQ(parent.new_actor_handles_.size(), 1);
+  ASSERT_EQ(parent.new_actor_handles_.back(), forkedHandle2.ActorHandleID());
+  parent.new_actor_handles_.clear();
 
   // Test serialization and deserialization for in-band fork.
   std::string buffer1;
@@ -811,12 +811,10 @@ TEST_F(ZeroNodeTest, TestActorHandle) {
   ASSERT_EQ(forkedHandle2.ActorCreationTaskFunctionDescriptor(),
             deserializedHandle1.ActorCreationTaskFunctionDescriptor());
   ASSERT_EQ(forkedHandle2.ActorCursor(), deserializedHandle1.ActorCursor());
-  ASSERT_EQ(forkedHandle2.TaskCounter(), deserializedHandle1.TaskCounter());
-  ASSERT_EQ(forkedHandle2.NumForks(), deserializedHandle1.NumForks());
 
   // Test out-of-band forking logic.
   ActorHandle forkedHandle3(parent, false);
-  ASSERT_EQ(2, parent.NumForks());
+  ASSERT_EQ(2, parent.num_forks_);
   ASSERT_EQ(parent.ActorID(), forkedHandle3.ActorID());
   ASSERT_NE(parent.ActorHandleID(), forkedHandle3.ActorHandleID());
   ASSERT_NE(forkedHandle2.ActorHandleID(), forkedHandle3.ActorHandleID());
@@ -824,11 +822,11 @@ TEST_F(ZeroNodeTest, TestActorHandle) {
   ASSERT_EQ(parent.ActorCreationTaskFunctionDescriptor(),
             forkedHandle3.ActorCreationTaskFunctionDescriptor());
   ASSERT_EQ(parent.ActorCursor(), forkedHandle3.ActorCursor());
-  ASSERT_EQ(0, forkedHandle3.TaskCounter());
-  ASSERT_EQ(0, forkedHandle3.NumForks());
-  ASSERT_EQ(parent.NewActorHandles().size(), 1);
-  ASSERT_NE(parent.NewActorHandles().back(), forkedHandle3.ActorHandleID());
-  parent.ClearNewActorHandles();
+  ASSERT_EQ(0, forkedHandle3.task_counter_);
+  ASSERT_EQ(0, forkedHandle3.num_forks_);
+  ASSERT_EQ(parent.new_actor_handles_.size(), 1);
+  ASSERT_NE(parent.new_actor_handles_.back(), forkedHandle3.ActorHandleID());
+  parent.new_actor_handles_.clear();
 
   // Test serialization and deserialization for out-of-band fork.
   std::string buffer2;
@@ -840,8 +838,6 @@ TEST_F(ZeroNodeTest, TestActorHandle) {
   ASSERT_EQ(forkedHandle3.ActorCreationTaskFunctionDescriptor(),
             deserializedHandle2.ActorCreationTaskFunctionDescriptor());
   ASSERT_EQ(forkedHandle3.ActorCursor(), deserializedHandle2.ActorCursor());
-  ASSERT_EQ(forkedHandle3.TaskCounter(), deserializedHandle2.TaskCounter());
-  ASSERT_EQ(forkedHandle3.NumForks(), deserializedHandle2.NumForks());
 }
 
 TEST_F(SingleNodeTest, TestMemoryStoreProvider) {
